Add Les_1_Task_3_формат вывода числовых переменных

+++ src/main/kotlin/Lesson_1/Task_4.kt
@@ -31,7 +31,7 @@ fun main(){
    val distance: Long = 40_868_600_000
     val age: Byte = 27
     val partOfDay: Float = 0.075F
-    val sec: Int = 6480
+    val sec: Short = 6480
     val partOfYear: Double  = 2.0547945205479453E-4
     val orbit: Int = 327000
